solved circular import bug that was rised by integration_test_building.py
