+++ data_generation/life_td_data_generation/utils/analysis/analysis.py
@@ -1,5 +1,3 @@
-from dataclasses import dataclass
-from enum import Enum
 
 import matplotlib.pyplot as plt
 import numpy as np  # arrays
@@ -9,6 +7,11 @@ from scipy.optimize import curve_fit
 from scipy.stats import norm
 from sdata import empty_dict
 from utils.analysis.finalplot import starcat_distribution_plot as final_plot
+from utils.analysis.histogram_utils import (
+    spectral_type_histogram,
+    SpectralType,
+    Plotparas,
+)
 
 # self created modules
 from utils.io import Path, stringtoobject
@@ -51,16 +54,6 @@ def fitfunction(name, x, y, arr):
 
 empty = empty_dict.copy()
 table_names = list(empty.keys())
-
-
-class SpectralType(Enum):
-    O = "O"
-    B = "B"
-    A = "A"
-    F = "F"
-    G = "G"
-    K = "K"
-    M = "M"
 
 
 def prepare_table(cat, columns):
@@ -168,40 +161,7 @@ def stellar_distance_histogram(arr, names, path, max_dist, xaxis):
     return
 
 
-def spectral_type_histogram(spectypes):
-    """
-    Makes a histogram of the spectral type distribution.
 
-    :param spectypes: array containing a spectral type for each star
-    :type spectypes:
-    :returns: array containig the number of stars for
-        each spectral type in spec.
-    :rtype: np.array
-    """
-    specdist = np.zeros(len(SpectralType))
-    for spectype in spectypes:
-        if spectype not in ["", "nan"]:
-            for j, spectraltype in enumerate(SpectralType):
-                if spectype[0] == spectraltype.name:
-                    specdist[j] += 1
-    specdist = specdist.astype(int)
-    return specdist
-
-
-@dataclass
-class Plotparas:
-    width = 0.15
-    color = ["tab:blue", "tab:cyan", "tab:green", "tab:olive"]
-
-
-def x_position(x, n_samples, sample_index):
-    """
-    Calculates x position for plots so that not all samples plotted over each other.
-    """
-    width_of_samples = n_samples * Plotparas.width
-    sample_location = sample_index * Plotparas.width
-    # if I change sign in front of sample_location result is that catalogs get shown in order backwards
-    return x - width_of_samples / 2 + sample_location
 
 
 def tight_plot(x, spec):
